2025-05-14-16.25/src/requirements.txt
Flask==2.2.3
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Werkzeug==2.2.3
SQLAlchemy==2.0.7
Pillow==9.5.0
PyMuPDF==1.21.1
python-docx==0.8.11
python-pptx==0.6.21
rarfile==4.0
bcrypt==4.0.1
email-validator==2.0.0
python-dotenv==1.0.0
gunicorn==20.1.0 
Flask-Mail==0.9.1 